0.5.7: improved 'convert' command

+++ cmd/convert.go
@@ -26,18 +26,18 @@ var convertCmd = &cobra.Command{
 func convert() {
 	configuration.load()
 	config := viper.GetString("config")
+	var path string
 	var err error
 	var data []byte
-	var path string
 	switch filepath.Ext(config) {
 	case ".json":
-		path = strings.Replace(viper.GetString("config"), "json", "yaml", 1)
+		path = strings.Replace(config, "json", "yaml", 1)
 		data, err = yaml.Marshal(&configuration)
-		fmt.Printf("Creating YAML file: %q\n", path)
+		fmt.Printf("Creating YAML file: %q...\n", path)
 	case ".yaml":
-		path := strings.Replace(viper.GetString("config"), "yaml", "json", 1)
+		path = strings.Replace(config, "yaml", "json", 1)
 		data, err = json.MarshalIndent(&configuration, "", "  ")
-		fmt.Printf("Creating JSON file: %q\n", path)
+		fmt.Printf("Creating JSON file: %q...\n", path)
 	default:
 		abortIfInvalidConfigurationFormat()
 	}
@@ -45,10 +45,14 @@ func convert() {
 		tools.PrintRedln("Configuration marshaling failed!")
 		log.Fatal(err)
 	}
+	if tools.PathExists(path) {
+		tools.PrintRedf("%q: file exists. Not overwritten.\n", path)
+		return
+	}
 	if err = ioutil.WriteFile(path, data, 0644); err != nil {
 		log.Fatal(err)
 	}
-	fmt.Println(">OK")
+	fmt.Println("> Ok")
 }
 
 func init() {
